// server/src/types/enums.ts
export enum Role {
  FLEET_MANAGER = 'FLEET_MANAGER',
  DRIVER = 'DRIVER',
  SAFETY_OFFICER = 'SAFETY_OFFICER',
  FINANCIAL_ANALYST = 'FINANCIAL_ANALYST',
}

export enum VehicleStatus {
  AVAILABLE = 'AVAILABLE',
  ON_TRIP = 'ON_TRIP',
  IN_SHOP = 'IN_SHOP',
  RETIRED = 'RETIRED',
}

export enum DriverStatus {
  AVAILABLE = 'AVAILABLE',
  ON_TRIP = 'ON_TRIP',
  OFF_DUTY = 'OFF_DUTY',
  SUSPENDED = 'SUSPENDED',
}

export enum TripStatus {
  DRAFT = 'DRAFT',
  DISPATCHED = 'DISPATCHED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}